+++ app/controllers/controle_acesso.py
@@ -10,7 +10,11 @@ def get_first_digit(number):
 
 @app.route('/controle_acesso')
 def controle_acesso():
-    data_atual = datetime.now().strftime('%d-%m-%y')
-    usuarios = [mongodb.get_registro_de_uso(data_atual)]
-    data = [{"nome": item["nome"], "tempo_de_uso": get_first_digit(item["tempo_de_uso"])} for item in usuarios]
+    try:
+        data_atual = datetime.now().strftime('%d-%m-%y')
+        usuarios = [mongodb.get_registro_de_uso(data_atual)]
+        data = [{"nome": item["nome"], "tempo_de_uso": get_first_digit(item["tempo_de_uso"])} for item in usuarios]
+    except Exception as err:
+        data = [{"nome":''}, {"tempo_de_uso": ''}]
+
     return render_template('controle_acesso.html', data=data)
